flatcar: Avoid building precompiled packages that don't work
Passing --kernel-name to nvidia-installer prevents the use of a
precompiled package. This argument is only needed when building for a
non-running kernel.

However, precompiled packages just don't work on Flatcar anyway, so
don't bother building them in the first place. The relinking step
somehow breaks the modules, but it is not yet clear why.

Signed-off-by: James Le Cuirot <jlecuirot@microsoft.com>